+++ wxdata/source/scwx/wsr88d/rda/volume_coverage_pattern_data.cpp
@@ -221,8 +221,7 @@ uint16_t VolumeCoveragePatternData::number_of_base_tilts() const
 double VolumeCoveragePatternData::elevation_angle(uint16_t e) const
 {
 
-   float elevationAngleConverted =
-      // NOLINTNEXTLINE This conversion is accurate
+   double elevationAngleConverted =
       p->elevationCuts_[e].elevationAngle_ * ANGLE_DATA_SCALE;
    // Any elevation above 90 degrees should be interpreted as a
    // negative angle
